get rid of warnings
* io.c (sysopen_func, rb_sysopen_internal): cast through VALUE to get
  rid of warnings.  fixup of r36355.
* process.c (rb_waitpid_blocking, rb_waitpid): ditto.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@36361 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/openssl/ossl_pkey_ec.c
@@ -1509,7 +1509,7 @@ static VALUE ossl_ec_point_mul(int argc, VALUE *argv, VALUE self)
 
         Check_Type(bn_v1, T_ARRAY);
         bignums_len = RARRAY_LEN(bn_v1);
-        bignums = (const BIGNUM **)OPENSSL_malloc(bignums_len * sizeof(BIGNUM *));
+        bignums = (const BIGNUM **)OPENSSL_malloc(bignums_len * (int)sizeof(BIGNUM *));
 
         for (i = 0; i < bignums_len; ++i) {
             bignums[i] = GetBNPtr(rb_ary_entry(bn_v1, i));
@@ -1522,7 +1522,7 @@ static VALUE ossl_ec_point_mul(int argc, VALUE *argv, VALUE self)
 
         rb_ary_unshift(points_v, self);
         points_len = RARRAY_LEN(points_v);
-        points = (const EC_POINT **)OPENSSL_malloc(points_len * sizeof(EC_POINT *));
+        points = (const EC_POINT **)OPENSSL_malloc(points_len * (int)sizeof(EC_POINT *));
 
         for (i = 0; i < points_len; ++i) {
             Get_EC_POINT(rb_ary_entry(points_v, i), points[i]);
